refactor(reditor): move directories
