+++ packages/Reditor/src/Note/index.ts
@@ -1,3 +1,4 @@
-export * from './model';
+export * from './line';
 export * from './note';
+export * from './types';
 export * from './useNoteOp';
